fix: delete pyi useless file
